requirements.txt
pytubefix>=8.7.0
pillow>=11.0.0
